New class has been added
A whole new class, named create(), has been added. From this class, users can generate the entire transition rate matrix associated with the queueing network specified in an adjacency matrix. In a future update it will be possible to use the create() object as input to the evaluate() class. Details will follow.